YEL-68 [chore] 오타 정리
