--- src/main/java/com/yello/server/domain/cooldown/response/RevealNameResponse.java
@@ -1,11 +1,0 @@
-package com.yello.server.domain.cooldown.response;
-
-import lombok.Builder;
-
-@Builder
-public record RevealNameResponse(
-    char name,
-    Integer nameIndex
-) {
-
-}
