--- src/main/java/com/yello/server/domain/question/service/QuestionService.java
@@ -1,7 +1,0 @@
-package com.yello.server.domain.question.service;
-
-import com.yello.server.domain.question.entity.Question;
-
-public interface QuestionService {
-    Question findByQuestionId(Long questionId);
-}
