+++ src/main/java/com/yello/server/domain/vote/common/WeightedRandomFactory.java
@@ -13,7 +13,7 @@ import static com.yello.server.global.common.util.ConstantUtil.THIRD_POINT_WEIGH
 import java.util.HashMap;
 import java.util.Map;
 
-public class WeightedRandom {
+public class WeightedRandomFactory {
 
     private static final Map<Integer, Double> weight = new HashMap<>();
