+++ src/main/java/com/yello/server/domain/vote/dto/response/RevealNameResponse.java
@@ -1,0 +1,17 @@
+package com.yello.server.domain.cooldown.response;
+
+import com.yello.server.domain.user.entity.User;
+import lombok.Builder;
+
+@Builder
+public record RevealNameResponse(
+        char name,
+        Integer nameIndex
+) {
+    public static RevealNameResponse of(User sender, int randomIndex) {
+        return RevealNameResponse.builder()
+                .name(sender.getName().charAt(randomIndex))
+                .nameIndex(randomIndex)
+                .build();
+    }
+}
