+++ src/main/java/com/yello/server/global/common/factory/ListFactory.java
@@ -4,9 +4,9 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
-public class ListUtil {
+public class ListFactory {
 
-    public static <T> List<T> toList(List<Optional<T>> optionalList) {
+    public static <T> List<T> toNonNullableList(List<Optional<T>> optionalList) {
         List<T> result = new ArrayList<>();
         optionalList.forEach(optional -> {
             optional.ifPresent(result::add);
