+++ src/main/java/com/yello/server/global/common/factory/PaginationFactory.java
@@ -1,14 +1,11 @@
 package com.yello.server.global.common.util;
 
+import org.springframework.data.domain.*;
+
 import java.util.Collections;
 import java.util.List;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.domain.Sort;
 
-public class PaginationUtil {
+public class PaginationFactory {
 
     private static final int PAGE_LIMIT = 10;
