+++ src/main/java/com/yello/server/global/common/util/ConstantUtil.java
@@ -19,7 +19,7 @@ public class ConstantUtil {
     public static final int NAME_HINT_POINT = 300;
     public static final int NAME_HINT_DEFAULT = -1;
     public static final int TIMER_MAX_TIME = 40;
-    public static final int TIMER_FIFTY_TIME = 40;
+    public static final int COOL_DOWN_TIME = 40;
     public static final int KEYWORD_HINT_POINT = 100;
 
 }
